Add optional label for metarun to child run edge
Summary:
This diff adds an optional `run_label` that can be set on the `meta_run` to
`run` association. This can be used to identify child runs within a metarun.

As a concrete example, distributed metaruns that include some number of
duplicate children can use this label to check for duplicates and only
upload the results once.

Differential Revision: D26010035

fbshipit-source-id: f5670376f05cd5e72fa91b8d985dc3897a364fdc