Remove --previous-input [2/2]
Summary: `--previousInput` is an obsolete parameter from when we used to pass the entire JSON output of the base analysis. It's been superseded by `--previousIssueHandles` for a couple years now, nothing except unit tests use it. I'll take anything I can get to make the CLI code a bit cleaner >.>

Reviewed By: fahndrich

Differential Revision: D26555204

fbshipit-source-id: fae221651fae5197e2ce8657eeee9fc2de97854f