add type intervals to metadata
Summary:
Since we now have the option to dump the class interval table to the zoncolan metadata file, we need some way to access it when we process the results. This diff adds an additional field to the metadata and adds a dict if the info is available.

Note we change the structure to create a dict based on the array and leave the entry out if the interval is null in the JSON.

Reviewed By: fahndrich

Differential Revision: D27477559

fbshipit-source-id: eb3c55841cc348415d7455b817142ad5e91e70d5