Introduce typed holders for issue data
Summary:
As data exits the parser and before it flows to the trace graph construction, convert them to strongly schematized types (NamedTuple).

This is makes the changes in the next diff easier to debug and understand. Also provides a good template for future tools on what they exactly must emit from their parser.

For now the conversion is after the data is already returned from the parser, but in the future it's possible to imagine pushing this schemification down into the various parsers.

Also probably provides a miniscule memory usage reduction as NamedTuples are more efficient storage wise than dicts. I didn't measure this diff in isolation, though, the total gains are measured in the next diff.
 ---
NOTE: This rebases on top of Maxime's landed diff D27586021 (https://github.com/facebook/sapp/commit/d8043c1922402aaa8cd6955ca2d67817c94971fc) which adds `TypedDict` types for the parser. I need `NamedTuple` types to realize the memory gains in the next diff, but doing that full conversion from `TypedDict` to `NamedTuple` everywhere would take me the entire rest of the week, so instead I define a parallel set of types (the same types I defined in V1 of this diff) and use those in the rest of the pipeline going forward.

Left a huge NB at the top of that file stating why copies of the types exist and what should eventually be done.

Reviewed By: fahndrich

Differential Revision: D27540208

fbshipit-source-id: ce4aff4f73caccc294613fdcdacb140409d5549b