Fix trimmed trace graph trace length computation
Summary:
Ibrahim noticed that some trace lengths were 0 where they shouldn't be.
https://fburl.com/zoncolan/hrtg1v1k

Turns out that we were filtering by source kinds in the initial frames, but this caused problems due to the taint-transform encodings. The first frames are only about the source/sink kinds of interest and don't need to be filtered.

Removed the extra filtering based on source/sink kinds.

Reviewed By: yuhshin-oss

Differential Revision: D27683921

fbshipit-source-id: 556b55edc4ad34a999db07fae4f9734dda8137e8