Highlights should show up from mariana trench in open source sapp view
Summary:
The highlights weren't showing up since we had an "empty" column range (1,1) and the `adjustRange` function was converting this to (-1, 0) which doesn't highlight anything.

After changing the `adjustRange` function, I got a highlight that was only one character long as expected, but this was a little hard to see. So I made it highlight the whole line to a more visible highlight

Reviewed By: dkgi

Differential Revision: D29117831

fbshipit-source-id: 43d40de204313b3b6753e76675b2e562739a1757