Added updating to saved filters (#22)
Summary:
This PR refers to https://github.com/facebook/sapp/issues/21 and adds the ability to update the filter field of a saved field from the web UI.

To do this, there’s just a couple of new changes in `filters.py` which inserts the new filter into the database or updates it (if the filter name already exists).

I first dug around a bit in the docs to see if there was anything like `UPSERT` available for SQLAlchemy, but I think many people online recommended to just query the database and update the fields manually (which seems to be alright, since there are only two fields that need to be updated in the object). I manually tested updating the filter fields manually to check that it works for all the filter options (codes, paths, callables, features, status, and trace lengths).

Please let me know if I missed something or if there’s anything that should be fixed/improved!

Pull Request resolved: https://github.com/facebook/sapp/pull/22

Reviewed By: arthaud

Differential Revision: D28964445

Pulled By: 0xedward

fbshipit-source-id: 28dc714a633923e1418cfddb28179c4d49d79f0b