Adds source and sink filtering ability to sapp
Adds ability to filter for issues in sapp with SourceName, SourceKind,
SinkName, and SinkKind options.

Tweaks the React UI to display the options in the FilterForm and
modifies the query that it makes with the backend flask server.

Implements the ability on the backend server to filter for issues based
on the given SinkName, SinkKind, SourceName, and SourceKind
filters/options.

Filtering by SourceName, SourceKind, SinkName, and SinkKind are via
additional graphQL subqueries that are concatenated and grouped just
like how features are filtered.

Modifies the HasAny IssuePredicate to have a single class for features
and the new mentioned options

Modifies tests to confront with the new filtering options.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>