Use ValueError as example exception
Summary:
Our linter complains if you use `assertRaises(Exception)`
because all exception types are subclasses, so it's very
easy for the test to not validate what you expect.

In this case it doesn't really matter, but I find lint
warnings distracting when I'm reading code so I switched
to ValueError instead.

Reviewed By: dkgi

Differential Revision: D29942807

fbshipit-source-id: 42581df53aa3972b5b2a0005bb6d39ff578ec93f