Mirror internal linter in open source
Mirror the tools used internally to check for linting errors in the open
source repository.

Modifies the lint action to run flake8, black, usort, and ESLint in
seperate instances.

Adds .flake8 file for flake8 mirroring that is used with pyre-check. Adds
.pyproject.toml file to configuration for usort. Modifies package.json
file in the react app to customize ESLint settings and adds a
.eslintignore file to specify folders to be ignored by ESLint.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>