Allow user to update a saved filter without clear
Currently, to update a saved filter, the user has to clear existing
filters, enter in the filter parameters, click save, and then enter the
existing filter name to update it.

Changed the workflow such that the save button isn't disabled when an
existing filter is selected and convey the filtername to the save model.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>