Allow user to update a saved filter without clear (#33)
Summary:
Currently, to update a saved filter, the user has to clear existing
filters, enter in the filter parameters, click save, and then enter the
existing filter name to update it.

Changed the workflow such that the save button isn't disabled when an
existing filter is selected and convey the filtername to the save model.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>
Fixes: https://github.com/MLH-Fellowship/pyre-check/issues/32

Pull Request resolved: https://github.com/facebook/sapp/pull/33

Reviewed By: dkgi

Differential Revision: D29989376

Pulled By: 0xedward

fbshipit-source-id: 34bddfa0b9ebbc6534c2ffb57bf1e779e2b2eaa6