Simplify development setup for sapp
Currently, we need to rebuild the react app each time when we make any
changes to the UI and hot-reload is unavialable. Changes that by
introducing hot-reload by running the app on 3000 port and the flask
server on port 5000 so that hot-reload is available for both the
frontend and backend.

Adds CORS headers for the same to enable communication between both
frontend and backend using flask-cors library while on debug
(development) mode. Also adds flask-cors to requirements.txt

Updates the docs to specify the same.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>