Support regex matching for source and sink filters
Adds support for regex matching for source name and kind and sink name
and kind filters. The regex matching occurs in the frontend and no
modification is hence required in the backend.

Adds new component and modifies callables which had regex matching to
also use the component as source and sinks do.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>