Suppress sqlite3.ProgrammingError (#37)
Summary:
Currently, sessions were created in threads and removed in the main
thread. This lead to sql alchemy complaining by emitting an
sqlite3.ProgrammingException. Moves local session removal to the same
thread in which it was created with the help of flask's teardown_request
callback and context stacks.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>
Fixes: https://github.com/MLH-Fellowship/pyre-check/issues/42

Pull Request resolved: https://github.com/facebook/sapp/pull/37

Reviewed By: arthaud

Differential Revision: D30102576

Pulled By: 0xedward

fbshipit-source-id: 9263bc7cd7ae3054eba3d4c12ebfe486f14bdcab