Create `queries.py`, extract leaves query
Summary:
We have a plan to build out an `extract` UI that's optimized
for the commandline and spits out json, but doing this in a
clean way might be tricky.

In the meantime, we want to allow a python script to grab
dropped-flow information from the sapp database so that we can
learn about where type coverage impacts our ability to get
static analysis results.

To do this without duplicating too much logic, this commit
creates a `queries` module where we can put raw query logic
that is useful both for extraction and for interactive
`explore` commands; since our use case only requires the query
from `leaves`, that's what we start with.

Reviewed By: arthaud

Differential Revision: D30118689

fbshipit-source-id: 9b369fac5e729dacce7305d04536cac89e4c5b34