Adds ability to import Filter(s) via UI (#34)
Summary:
Adds feature to import one or more filter json files to the database via
the react UI.

Adds a draggable component through which filter json files can be
dragged into or selected with the help of the on system file picker.
Only json files are alllowed and the other sort of files are skipped.
File reading is done on the client side, in the react UI, the contents
are then passed to the graphQL database via the existing save filter
mutation.

Modifies the mutation associated function so that they won't load the
last imported filter but retain such functionality on saving a new
filter as before.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>

Pull Request resolved: https://github.com/facebook/sapp/pull/34

Reviewed By: dkgi

Differential Revision: D30020999

Pulled By: 0xedward

fbshipit-source-id: 16432f527c284430cea81a601bd1d126ebcb02cb