Fix failing ui tests (#45)
Summary:
Jest tests were failing because the apollo mock provider was returning
an invalid response (a new attribute was added to issues following PR#43).
Also modifies the snapshot to confront to updates.

ESLint was complaining of an unused import. Readded export option in the filter
which was lost somehow (rebase?)

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>

Pull Request resolved: https://github.com/facebook/sapp/pull/45

Reviewed By: dkgi

Differential Revision: D30600556

Pulled By: 0xedward

fbshipit-source-id: 730de60bd64f959bb49d3be87c0ecc3b3e7e92cc