Add syntax highlighting support for more languages
Previously syntax highlighting of code in traces section was confined to
just python code using CodeMirror. Since, SAPP is not just confined to
analyze results of python based static analyzers, and it is capable of
supporting other static analyzers, adds support for many more languages
including Java, Kotlin, C++, C, C#, etc. while retaining support for
python.

CodeMirror requires "code-mode" js files to be imported for it to
support that language via the mode parameter. Support for C and other
languages in addition to Java was added since they are present in the
c-like.js mode file needed for Java.

Decision to decide which language mode to use is based on file
extension with the help of a defined javascript object. Fallback if a
unsupported extension is encountered is python.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>