Makes Dalvik Code in source and sinks readable
Sometimes, sink names and source names in issues can take up the form of
Dalvik Codes which are hard to read. Adds code to make use of the
existing Davlik to Human Readable converter used for readables apply to
sink and source names as well.

Updates jest tests to reflect the same.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>