Recompute filtered issue paths
Summary:
This diff finally fixes a problem with trace lengths on diffs.

Because we filter traces by affected files, we may be getting rid of some short trace that originally led to an issue. In that case, the trace lengths on the issue is too short and might pass a filter it shouldn't.

We now perform a full trace traversal (up to a max length of 100) of the issues identified after filtering to affected files and recompute the length to the sink/source. We update the frames' leaf assoc lengths so trace expansion should get the shorted correct path easily.

Reviewed By: williewillus

Differential Revision: D31945656

fbshipit-source-id: 14d83d966c5c3320f35e652ffb0ae2fed50ec696