Fix line-number off by one error in MT parser
Fix line-number being off by one in MT parser.
MT for some reason, reports line number as 0 which can lead to issues in
the frontend. Add one to it to prevent this problem.

This can also be something that is wrong with Mariana Trench as the
general convention is to report line numbers from 1 rather than 0.
Hence this commit is pretty open to debate.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>