Set pool recycle time lower
Summary:
We are seeing post-processing timeouts during writing of leaf-assocs (see task). The failure suggests that we might be starving the pool of connections.

I checked the pool_recycle settings as suggested here https://www.internalfb.com/intern/wiki/MySQL/MySQL_Gateway/ClientLibraries/Common_Problems/#idle-wait-timeout

The FB default is 9s. In the context where we are failing, we tried to write 14.5M leaf assocs in 30k bundles. Each bundle is it's own connection/transaction. Thus, we have 484 bundles - thus connections.
A successful run finishes a bundle every 0.75-0.8 secs. Thus, we can calculate how many connections we need simultaneously (pool_recycle/ 0.75), which is about 12 with the default 9s timeout. I lowered the timeout to 5s, so we don't need that many connections.

For good measure, also increased the connection timeout as we do need time to dedup stuff and with more data, this time goes up.

Reviewed By: alambert

Differential Revision: D33025308

fbshipit-source-id: 833521f13e95d3ef4c1555a7336ade026c89debd