Match GraphQL CORS header with frontend hostname and port in debug mode
`start.js` allows users to specify the host and port to serve the frontend from
by relying on the `HOST` and `PORT` environment variables. If these environment
variables are not set, the frontend resorts to a default host `0.0.0.0` and port
`3000`.

    const DEFAULT_PORT = parseInt(process.env.PORT, 10) || 3000;
    const HOST = process.env.HOST || '0.0.0.0';

The way the frontend extracts the data is by querying the `/graphql` endpoint
which sets a CORS header in debug mode via `server.py`.

A bug was discovered when users set the `HOST` and/or `PORT` environment
variable to something other than `0.0.0.0` and `3000` respectively. The CORS
header for the `/graphql` endpoint would prevent the frontend from being able to
query it.

This commit introduces a solution by matching the code for specifying the
frontend host and port in the CORS header set by `server.py`.

    -        CORS(application, resources={r"/graphql": {"origins": "http://localhost:3000"}})
    +        default_port = int(os.environ.get("PORT"), 10) or 3000
    +        host = os.environ.get("HOST") or "localhost"
    +        CORS(
    +            application,
    +            resources={
    +                r"/graphql": dict(origins="http://" + host + ":" + str(default_port))
    +            },
    +        )

My only concern with this current implementation is that the behaviour of
`parseInt()` in JavaScript may differ from Python's `int()` function for
specific edge cases and may result in unexpected issues.

Additionally, the `README.md` file was updated to include a warning concerning
the reliance on the two aforementioned environment variables.