Fix merge session scope
Summary:
Bulk save sometimes produces errors that a SQL connection has timed out.

This seems to be due to the fact that we open a session and then spend a lot of time preparing items to save (sorting deduping) without touching the session.

Solution:
When preparing to save, don't open a session, as it is only used for SharedTexts and Issues.

Instead pass the database to the merge overrides, so sessions are only created when we need it.

In the merge_keys case, we open the session once per bucket instead of overall once.

Reviewed By: compositor

Differential Revision: D33537287

fbshipit-source-id: 1376f9f496e91b5d8d4aa0c6a79cf15da09d157d