Match GraphQL CORS header with frontend hostname and port in debug mode (#66)
Summary:
**Pre-submission checklist**
- [x] I've ran the following linters locally and fixed lint errors related to the files I modified in this PR
    - [x] `black .`
    - [x] `usort format .`
    - [x] `flake8`
- [x] I've installed dev dependencies `pip install -r requirements-dev.txt` and completed the following:
    - [x] I've ran tests with `./scripts/run-tests.sh` and made sure all tests are passing

`start.js` allows users to specify the host and port to serve the frontend from by relying on the `HOST` and `PORT` environment variables. If these environment variables are not set, the frontend resorts to a default host `0.0.0.0` and port `3000`.

```js
const DEFAULT_PORT = parseInt(process.env.PORT, 10) || 3000;
const HOST = process.env.HOST || '0.0.0.0';
```

The way the frontend extracts the data is by querying the `/graphql` endpoint which sets a CORS header in debug mode via `server.py`.

A bug was discovered when users set the `HOST` and/or `PORT` environment variable to something other than `0.0.0.0` and `3000` respectively. The CORS header for the `/graphql` endpoint would prevent the frontend from being able to query it.

This Pull Request introduces a solution by matching the code for specifying the frontend host and port in the CORS header set by `server.py`.

```diff
-        CORS(application, resources={r"/graphql": {"origins": "http://localhost:3000"}})
+        default_port = int(os.environ.get("PORT"), 10) or 3000
+        host = os.environ.get("HOST") or "localhost"
+        CORS(
+            application,
+            resources={
+                r"/graphql": dict(origins="http://" + host + ":" + str(default_port))
+            },
+        )
```

My only concern with this current implementation is that the behaviour of `parseInt()` in JavaScript may differ from Python's `int()` function for specific edge cases and may result in unexpected issues.

Additionally, the `README.md` file was updated to include a warning concerning the reliance on the two aforementioned environment variables.

Pull Request resolved: https://github.com/facebook/sapp/pull/66

Test Plan:
The following steps should allow one to reproduce the behaviour described above in the `main` branch.

```
$ export PORT=8080
$ python3 -m sapp.cli server --debug
$ cd sapp/ui/frontend && npm run-script start
```

![image](https://user-images.githubusercontent.com/18099289/145718073-f9feb1fd-3471-42d7-96ab-e4e501d4b88f.png)

The changes introduced by this PR should resolve the test case above and no CORS errors should be encountered.

Reviewed By: gbleaney

Differential Revision: D33197946

Pulled By: 0xedward

fbshipit-source-id: 18cc2ff73816568bb5311e494571fe93d6307a09