fix contents index length
Summary:
When creating our mysql dbs, we seem to miss an index length on the ix_messages_handle table. It seems that total index length has to be <=3072, but since we have the string + the kind, we need to reduce it.

I think this is behavior change from sql 5.6 where it seemed to default to 767 (at least that is what we currently have in our aosc schema). So I'm matching that.

repro:
  % ~/fbcode/buck-out/gen/zoncolan/tools/zoncolan-drop-table.par
  % ~/fbcode/buck-out/gen/zoncolan/bin/zoncolan-cli.par setup --database xdb      --dbname devdb.$USER --repo-dir dummy

Note, the error only shows the first time you run setup after dropping the tables. Afterwards, it isn't trying to create the index again.

Reviewed By: williewillus

Differential Revision: D34171648

fbshipit-source-id: e4b2055fec377343ba2341d3a845a15caee780b9