Use the new issue handle
Summary:
# Problem
To identify unique issues across runs, sapp uses a issue handle composed of the root callable, the rule code, the relative line number and column number where sources meets sinks (more precisely, the position of the sink). The problem is that a small line change between two runs will triggers different issue handles, and hence different issues, leading to a lot of duplicates.

# Solution
Instead of using the line number and column number, let's use the name of the callee leading to the sink, the argument name, as well as the textual order index of the callee in the call graph.
This is robust to line changes. The downside is that there is a small risk of ignoring new issues when adding a call to a function before an existing call to the same function, as this will mess up indices.
For instance:
```
def f():
   a = source_with_hop()
   sink_with_hop(x=a) # textual order index is 0
   sink_with_hop(y=a) # textual order index is 1
   sink_with_hop(z=a) # textual order index is 2
```

# Implementation
The new issue handle has been implemented in Pysa and is available in the `master_handle` attribute of an issue. Here, we just need to read that value when available.

Reviewed By: alexkassil

Differential Revision: D34541509

fbshipit-source-id: 449e7d2766ef118d4674a08c41f1e36a186a2636