Open source the json command
Summary:
This diff moves the `json` command to open source. This can be used with `sapp json` or just `json` within `sapp explore`.
This command allows one to pretty print the json model and issues for a given callable, directly from the analysis output. This is useful for debugging false negatives, since sapp does not store the full models in the database.

Reviewed By: dkgi

Differential Revision: D34806858

fbshipit-source-id: 7c451d511343e83b8268d1549a0485c4fa36168a