Properly invalidate the lookup table when looking at a different analysis output
Summary:
When using the `json` command to explore analysis output models, we compute and cache a lookup table. This lookup table is basically a mapping from callables to the offset of their model and issues in a file.
Unfortunately, this lookup table is never invalidated, so when looking at different runs, it becomes invalid.
This diff adds a simple invalidation check based on the location of the files.

Reviewed By: dkgi

Differential Revision: D34808845

fbshipit-source-id: 7c88c82e745bcb6e4e4c106254450e0dce448c94