Add an option to discard issues that already exist in the given metarun
Summary:
# Problem
When performing meta runs with many runs analyzing overlapping source files (e.g, dependencies), there might be a big amount of issue duplicates, which drastically increases the size of the database.

# Solution
As a solution to reduce the database size, this diff adds an option that discards issue instances that already exist in the current metarun. Note that since issue equality is very costly (in theory, we should check all paths), this diff uses the following heuristic: two issues are identical if they have the root callable, rule, root location, source leaves, sink leaves, source kinds, sink kinds, breadcrumbs, and minimum distance to source and sink leaves. This makes it possibly to compute a hash and efficiently check for the existence of an issue, using an index on a table. The downside is that we might discard issues with slightly different paths.

Reviewed By: fahndrich, alexkassil

Differential Revision: D39118916

fbshipit-source-id: ee1ca2547ca444e0ba39ac9da832e1e60896da89