Quick fix for #89: Bump pyre-extensions version in requirements.txt (#90)
Summary:
Bump pyre-extensions version in requirements.txt
Since, these are extensions are for the typing module, I think we can remove the version information as well.

**Pre-submission checklist**
- [x] I've ran the following linters locally and fixed lint errors related to the files I modified in this PR
    - [x] `black .`
    - [x] `usort format .`
    - [x] `flake8`
- [x] I've installed dev dependencies `pip install -r requirements-dev.txt` and completed the following:
    - [x] I've ran tests with `./scripts/run-tests.sh` and made sure all tests are passing

A simple version bump for pyre-extensions.
In order to prevent dependency issues in sapp, we decided to pin versions in the requirements.txt. Now a package that pyre-check introduces a dependency conflict as it depends on a newer version of a package (pyre-extensions)
Bumping pyre-extensions which extend the typing module. Since it is non-breaking module (pyre-extensions), we could potentially unpin the versioning information and allow pyre-check to dictate the version. However, I could be wrong here.

Pull Request resolved: https://github.com/facebook/sapp/pull/90

Test Plan:
> Ran 102 tests in 5.983s

The frontend tests are failing however, but I strongly believe them to be un-related to this PR.

Fixes https://github.com/facebook/sapp/issues/89

Reviewed By: alexkassil

Differential Revision: D43845657

Pulled By: stroxler

fbshipit-source-id: 9c325114ab9d602b81b9fe88240b2cc12b9cf3b6