Fix TestSappCli (again)
Summary:
TestSappCli is [failing](https://www.internalfb.com/intern/test/562950040173828?ref_report_id=0) again with a similar errors to D41978510 and D42854436.

Considering how this is only failing on line 80 as opposed to the line 77, I believe the issue is how we're indexing the temporary path. In line 80, we assume that the directory only has three letters, which it may very well not. This diff removes the indexing and now we can properly check whether the temp path is in the path for the linemap.

Reviewed By: arthaud

Differential Revision: D43959544

fbshipit-source-id: 68e8b8c84f5cdeb9912dd5717b8bcf15738454ca