start logging RSS before/after parsing
Summary: Context: Looking into improving the amount of memory we use during parsing for some internal, large usecases. Can't improve what you don't measure, so let's start tracking some basic RSS during parsing, which seems to be where SAPP is spending most of it's time, and use this as a guide of whether our changes are having the expected impact.

Reviewed By: alambert

Differential Revision: D44599026

fbshipit-source-id: 998632c9093a139987db9219a675befecb7207f1