JSON parser: Parse updated JSON entries for local features/positions
Summary:
JSON entries for local features/positions are being moved out of "kinds" into the outer "taint" structure (see updated test cases for examples).

The parser is updated to handle this input. Note that this new parser does not work on the old input (it would miss out on local positions/features but should not fail).

There's a subtle issue with the `normalize*` methods: We were re-constructing the taint object assuming its keys were only "call" and "kinds". This would cause the new "local_positions" and "local_features" fields to be dropped. This diff changes it to copy the original object, preserving all fields, and only overwrite/drop fields that need to be "normalized".

Reviewed By: anwesht

Differential Revision: D44977963

fbshipit-source-id: f01e203b28f3507b622bc3b97b9d818f0c68bcb8