Retry batch inserts on duplicate key exceptions
Summary:
This diff fixes a race condition where another run simultaneously creates an issue with the same handle. This could causing the script to crash with a "Duplicate entry for key 'issues.handle'" exception from the database.

This race happened when the duplicate issue was inserted after we checked for existing issues and before we inserted new issues.

I chose to fix this by retrying the batch insertion with smaller batches until those batches succeed or we get to the point where the failed batch is only one record and then "re-merging" that record using the logic to check for existing issues.

I wan't able to find a way to tell which objects in the batch failed or to tell SQLAlchemy to upload the batch but ignore duplicate key failures (ie: generate an `INSERT IGNORE` statement in SQL), so I believe this binary-search-like approach is reasonable.

Reviewed By: yaabdalla

Differential Revision: D45999992

fbshipit-source-id: 02a095f9036754cf67f4ff22f7847520cd365507