Stricter typing and state checks for `DBID`
Summary:
While working on D45999992, I was somewhat surprised to see that `DBID` allows us to set the inner id multiple times, even if the id was already retrieved out of the object.

While we do need the ability to set the `id` multiple times, it seems prudent to add a check that ensures that we don't one value, read that value, and then set another value later.

This diff adds checks for that case as well as well as some type annotations to `DBID`.

Reviewed By: fahndrich

Differential Revision: D46088586

fbshipit-source-id: 75c9d22336e1b057b6e02fe485ff0aa673f8239e