Mariana Trench Parser: Parse propagation frames
Summary: Now that we are tracking "next hop" traces for propagation taint trees, we need to parse and write these to sapp for the subtraces. This diff now consumes "propagation" keys the same way as "sources" and "sinks". The frames corresponding to regular propagations without traces are filtered out. When writing to sapp, the "propagation" traces are considered to be preconditions (i.e. sink traces).

Reviewed By: arthaud

Differential Revision: D48018884

fbshipit-source-id: ab367e29f30581861a07f678809f31c70271cbea