Parser: Support new "taint_origins" field in json
Summary:
Parse the new "taint_origins" field. This assumes "origins" is the key in the json, which is not currently the case (but will be in the next MT diff).

TODO: Populate "taint_origins" correctly for CRTEX leaves, and remove some special-handling around CRTEX. Note that `normalize_crtex_conditions` still needs to happen (unfortunately)

Reviewed By: anwesht

Differential Revision: D50046193

fbshipit-source-id: 9962964d1e7f3f16fd276fca934985a3512e9091