Use instance variables in `PrimaryKeyGenerator` and add tests
Summary:
While writing tests for `PrimaryKeyGenerator`, I saw that the `pks` variable being initialized as a class variable instead of an instance variable. Although scripts typically only have one instance of the key generator, we do not want two generator instances to share state like this. For this reason, `pks` is now an instance variable.

For similar reasons, I also moved the `PRIMARY_KEY` and `QUERY_CLASSES` class variables so that they need to be explicitly set in the constructor of the key generator and are not shared globally.

Finally, this diff adds some tests to exercise the behavior of the PrimaryKeyGenerator to make sure I don't regress anything with some upcoming changes.

Reviewed By: fahndrich

Differential Revision: D50905162

fbshipit-source-id: c2b95b157a491a6ce66333cd1fca07742225a23a