Add ability for `PrimaryKeyGenerator` to only generate an allowlisted range of IDs
Summary:
To support database sharding where IDs are unique to a specific database, allow setting an `allowed_id_range` to a custom range of integers.

When this is set, the `PrimaryKeyGenerator` will take care to only ever vend IDs from that range. If this range becomes exhausted, id reservation will fail.

For additional safety, the key generator will **not** automatically ratchet ids forward so they are within the provided range: if it looks like existing data has already been written outside the allowed range, id reservation will fail. This could be a sign that the range is misconfigured.

Reviewed By: fahndrich

Differential Revision: D50856836

fbshipit-source-id: e8607a2675c183f50fa223d6c9784fc8ad514c97