Revert D23522664 so we can upgrade create-react-app
Summary:
# Problem
`npm install && npm run build` currently returns errors because of outdated npm packages.

# Solution
Instead of trying to manually upgrade everything, which would be a huge amount of work, let's try to upgrade "create-react-app", which is the tool we used to bootstrap the UI.
Unfortunately, we opted out of "create-react-app" by using "npm run eject", which means we wouldn't follow upgrades.
This diff reverts that change so we can easily upgrade in a future diff.
This was done with `patch -R` and then manually fixing a few conflicts.

Reviewed By: alambert

Differential Revision: D51232344

fbshipit-source-id: 83ee1bb4eb2ae0f7f2f170b52d074c4c53f3b658