Disallow Method.from_json(None). Removing concept of "leaf" method
Summary:
Currently, `Method.from_json(None)` returns `Method("leaf")` which is strange. If a method doesn't exist, it doesn't exist. "Leaf" is not a method.

This diff replaces "leaf" methods with `Optional[Method]`. In most cases, the method (usually the callee) should not be `None` anyway, otherwise the analysis is emitting a bad json.

Reviewed By: anwesht

Differential Revision: D51187701

fbshipit-source-id: 8f5a0b92b3e1475b5751612ddaad961e32bad8c2