More efficient handling of races to write duplicate Issues
Summary:
Previously, D45999992 added retries on duplicate key exceptions to fix a race condition where another run simultaneously creates an issue with the same handle.

Unfortunately, this approach falls apart in high-contention situations where there are many duplicate issue handles. When multiple processes are fighting to create the same set of Issue handles, this can create pathological behavior where we send even more `INSERT` statements than the number of Issues in a batch.

To fix this poor performance, we remove the retry logic and instead leverage the ability for most database engines to ignore key conflicts and continue inserting rows that do not conflict. SQLAlchemy doesn't provide a database agnostic way of doing this, so we implement it for MySQL and SQLite.

After doing this insert, we re-read the records from the database and reassign any IDs for records that some other process ended up writing.

Reviewed By: fahndrich

Differential Revision: D51688817

fbshipit-source-id: 98e47fac009b255960dbbb413670ad31c4573f29