Improve performance of `_merge_by_keys`
Summary:
While working on D51688817, I noticed that the "prepare" phase for SharedTexts took longer than actually saving those SharedTexts. This diff attempts to improve performance of `_merge_by_keys` to speed up the prepare phase in this case.

To make this happen, we do two things:
* `_merge_by_keys` now uses a `SELECT ... WHERE (...) IN (...)` statement rather than a `SELECT ... WHERE (...AND...) OR (...AND...) OR (...AND...)` statement. This is not only shorter but seems to perform better when looking up large numbers of records
* The batch size of each `SELECT` statement is now much larger to reduce the number of round-trips we need to make checking existing records

As part of this change, we also remove the `hash_item` lambda argument. This appeared to be a workaround for the non-hashability of lists, but we can simply use a tuple instead. I was also concerned about accidental collisions if two items with different keys produced the same hash string, so that's no longer a possibility.

Reviewed By: arthaud

Differential Revision: D51730133

fbshipit-source-id: aa7ff68982f5dd0bb6140ebf5ad2d93eec8177fc