Parser: Support class intervals
Summary:
Support parsing class intervals.

In the analysis, intervals are represented as: LocalTaint -> Kind -> Interval -> Frame

In the json output, it is: LocalTaint -> Frame, where Frame contains the kind and interval. It's not as nice from a "reflect the underlying representation" perspective, but it reduces nesting and improves readability. I'm inclined to stick with the current representation.

In SAPP, a trace_frame is uniquely identified by "caller, callee, caller port, callee port, interval". Therefore, the parser has to do some interval deduplication (see `Kind.partition_by_interval(...)`). A similar deduplication logic already exists for callees in origin frames. This diff adds "intervals" to the mix.

Reviewed By: anwesht

Differential Revision: D51931277

fbshipit-source-id: 75a8629682dc41e5884287e0eb29807e21aad438