Parser: Prefix origin callee ports with "[sink|source]:<actual port>"
Summary:
Remove the whole "leaf" thing in `Port.from_json(...)` which used to determine when a port should be "source" or "sink".

Instead, `Origin.from_json` (which is where this "leaf" thing is used) will determine what the port should be, i.e. specifically prefix it with "source:" or "sink:" which SAPP needs for traces to work, then append the actual port.

Reviewed By: anwesht

Differential Revision: D55001036

fbshipit-source-id: 8fd14c40a5497c9abf746061d8c1ae30a93796e7