Parse frame_type for extra_traces
Summary:
This diff updates the mariana trench sapp parser to read "frame_type" from the json instead of defaulting to "tito_transform".
Note: This introduces a backwards incompatible change as we now expect the "frame_type" key to exist within an extra-trace json. This should not be a problem as we release the binary and `mariana_trench.tools` fbpkg with the zoncolan-cli together. Expect to fail if running on results from the older analysis binary.

Reviewed By: yuhshin-oss

Differential Revision: D57893393

fbshipit-source-id: 02143ba2da51018286b07aaf0d283a208d8fb452