update JSON representation of TransformKind
Summary:
This diff changes the JSON representation of `TransformKind` from a string to a
JSON object that spearates the global / local transforms as well as the base
kind, which makes parsing easier.

The new JSON object has the following format:

```
{
  "base" : <base_kind_str>,
  "global" : <global_transformation_str>,
  "local" : <local_transformation_str>
}
```
as a reference, the old string-based format is:
```
<local_transformation_str>@<global_transformation_str>:<base_kind_str>
```

The reason we still use strings rather than JSON arrays for global and local
transformations is that the MT integration tests tend to sort the JSON arrays
and break the order of transforms.

Reviewed By: yuhshin-oss

Differential Revision: D57931089

fbshipit-source-id: 90ebc5f475da0b6aec933eb23f6f4d0cb5bf1ffa