Add a SourceLocationWithFilename type to parse locations
Summary:
# Problem
Right now, we use `ParsePosition` to parse and store locations with a filename.
Unfortunately, `ParsePosition` is implemented with a `TypeDict` that is non-total. It has two main drawbacks:
* It is not hashable (which we will need in a follow-up)
* It might have missing keys, for instance `filename` might not be defined, it won't complain

# Solution
Let's implement our own type `FullSourceLocation`, which is basically a `SourceLocation` with a filename.
We also use that opportunity to clean up the parser and have `_parse_location` and `_parse_full_location` methods.

Reviewed By: tianhan0

Differential Revision: D58528699

fbshipit-source-id: 298fd78f302bf0d0e2a2e92297cb07893cd133b9