Emit a single frame per callee, callee port and location
Summary:
In the future (see D58465386, D58468250), we will want to emit different trace frames if the breadcrumbs are different. Right now, we only emit a single trace frame for all kinds under a local taint.

To prepare for that change, this diff adds logic to merge and deduplicate trace frames with the same "key". Right now, the key is defined as the callee, callee port, callee location and type interval.

This should technically be a no-op since pysa only emit a single frame per "key" right now.

Reviewed By: tianhan0

Differential Revision: D58582725

fbshipit-source-id: 05146afecfcf0e6dcf77ec1278c1e60a8ceebc69