Update Issue representation to hold an exploitability origin
Summary:
This diff adds a std::variant to represent 2 types of Issue callees:
1. `string` as before
2. `ExploitabilityOrigin`: The issues will be grouped by this.

When creating an issue with the `ExploitabilityOrigin` as the "callee", we add the position of the Method being analyzed as the Issue's position as we could have multiple call-sites in the current method that leads to the same exploitability-origin, and hence must be grouped together. Also, the `Issue` frame in sapp only supports a single line, so no point tracked a set of positions either.

Reviewed By: arthaud

Differential Revision: D62909982

fbshipit-source-id: deabf70376c77477d2fab50c5d02f577905313ed