Add ability to spread to frames not matching callee port and location
Summary:
In some cases, for example, if the callee port and location doesn't match and the context providing frame is a partial flow that doesn't neccessarily intercept to the context recieving flow, let's still add the option to add the context breadcrumb to the trace frame.

This allows this feature to be re-used as adding a context feature to a callable when the callee ids and caller ids match.

The following flow was modelled here:

Crypto Init -> genEncrypt/Decrypt

This flow will have a via-crypto-project-value breadcrumb with the name of the crypto project.

This will be added to the two following flows:
1. User Controlled -> genDecrypt -> ReturnedToUser
2. User Controlled -> genEncrypt -> ReturnedToUser

If we model the first context flow as CryptoInit -> genEcncrypt -> ReturnedToUser, we add the breadcrumb to many trace frames which will add incorrect crypto project names if the flows (1 and 2) have some common trace frames among themselves.

With this feature, that is dealt with.

However, we still lack the capability to add the breadcrumb from trace frame to issue instance.

Reviewed By: fahndrich

Differential Revision: D62130425

fbshipit-source-id: c313764ac43d48f18f862277ed341c3877dc8f15