Strict typing for Summary object
Summary: The `Summary` object carries run information through the SAPP pipeline. It was typed as a dict of string to heterogeneous object where the keys are hardcoded throughout the code. This can be improved significantly with [`TypedDict`](https://docs.python.org/3/library/typing.html#typing.TypedDict) so that the type checker knows which keys are valid and what the types of each of their values are.

Reviewed By: fahndrich

Differential Revision: D66671979

fbshipit-source-id: e9149162c2aca774689bafcb878480589205a9ff