Temporary workaround for missing type information in Flask
Summary:
A bug in Pyre currently leads to missing type information when using the `aparrapo.route` decorator (from the `Flask` class).
The bug is due to a fundamental design flaw in Pyre, which will hopefully be resolved in Pyre2, which should be out in a half or two.
In the meantime, we can use a workaround by hiding the `__getattr__` function under a `if not TYPE_CHECKING:`. This `__getattr__` function hides the whole `flask.sansio` module from Pyre, which leads to a bug when resolving decorators. See T212388542 for more information.

Reviewed By: tianhan0

Differential Revision: D68211028

fbshipit-source-id: c601131601714f3045313dcd0ff0b11189f4614b