Add upward propagated breadcrumbs to main traces
Summary:
This diff adds upward propagated breadcrumbs to main traces in SAPP.

This will allow SAPP to find traces with the upward propagated breadcrumb even if the breadcrumb originates on a subtrace . This is the same technique for tool inferred and propagated breadcrumbs from subtraces.

Note: Eventually, we will want to avoid this inlining of breadcrumbs by tools and SAPP and search the full subtraces.

Reviewed By: abishekvashok

Differential Revision: D69793447

fbshipit-source-id: 65690ef5e335ea4d77c6bd4e988b4549bd1415f2